Remove `BaseSecurityManager`
